espresso_api::proto::explorer_service_server

Trait ExplorerService

pub trait ExplorerService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_explorer_block_detail<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetExplorerBlockDetailRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerBlockDetailResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_explorer_block_summaries<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetExplorerBlockSummariesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerBlockSummariesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_explorer_transaction_detail<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetExplorerTransactionDetailRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerTransactionDetailResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_explorer_transaction_summaries<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetExplorerTransactionSummariesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerTransactionSummariesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_explorer_summary<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetExplorerSummaryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerSummaryResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_explorer_search<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetExplorerSearchRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerSearchResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ExplorerServiceServer.

Required Methods

fn get_explorer_block_detail<'life0, 'async_trait>( &'life0 self, request: Request<GetExplorerBlockDetailRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerBlockDetailResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get one block in full

fn get_explorer_block_summaries<'life0, 'async_trait>( &'life0 self, request: Request<GetExplorerBlockSummariesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerBlockSummariesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get a window of blocks ending at the target block

fn get_explorer_transaction_detail<'life0, 'async_trait>( &'life0 self, request: Request<GetExplorerTransactionDetailRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerTransactionDetailResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get one transaction in full, with its payloads

fn get_explorer_transaction_summaries<'life0, 'async_trait>( &'life0 self, request: Request<GetExplorerTransactionSummariesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerTransactionSummariesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get a window of transactions ending at the target transaction

fn get_explorer_summary<'life0, 'async_trait>( &'life0 self, request: Request<GetExplorerSummaryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerSummaryResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get an at-a-glance snapshot of the chain

fn get_explorer_search<'life0, 'async_trait>( &'life0 self, request: Request<GetExplorerSearchRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ExplorerSearchResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Search the chain for blocks and transactions matching a hash

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".