espresso_api::proto::token_service_server

Trait TokenService

pub trait TokenService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_total_minted_supply<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTotalMintedSupplyRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<TotalMintedSupplyResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_circulating_supply<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetCirculatingSupplyRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CirculatingSupplyResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_circulating_supply_ethereum<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetCirculatingSupplyEthereumRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CirculatingSupplyEthereumResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_total_issued_supply<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTotalIssuedSupplyRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<TotalIssuedSupplyResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_total_reward_distributed<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTotalRewardDistributedRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<TotalRewardDistributedResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with TokenServiceServer.

Required Methods

fn get_total_minted_supply<'life0, 'async_trait>( &'life0 self, request: Request<GetTotalMintedSupplyRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<TotalMintedSupplyResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the total supply of the ESP token minted on Ethereum; excludes unclaimed rewards

fn get_circulating_supply<'life0, 'async_trait>( &'life0 self, request: Request<GetCirculatingSupplyRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CirculatingSupplyResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the circulating supply, following the mainnet unlock schedule

fn get_circulating_supply_ethereum<'life0, 'async_trait>( &'life0 self, request: Request<GetCirculatingSupplyEthereumRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CirculatingSupplyEthereumResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the circulating supply of ESP tokens on Ethereum L1

fn get_total_issued_supply<'life0, 'async_trait>( &'life0 self, request: Request<GetTotalIssuedSupplyRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<TotalIssuedSupplyResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the total issued supply, including rewards not yet claimed on Ethereum

fn get_total_reward_distributed<'life0, 'async_trait>( &'life0 self, request: Request<GetTotalRewardDistributedRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<TotalRewardDistributedResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the total rewards distributed by consensus, including rewards not yet claimed on Ethereum